Moved 2024 "Views" into a subfolder, for consistency. (#235)
* Moved 2024 "Views" into a subfolder, for consistency.

* Minor fixes to privacy policy text, plus remove "Get Tickets" here.
